Remove datasheet from repository
We don't want these as this probably brings up all sorts of licensing
nonsense, and adds unnecessary bulk to the repository. Removing it from the
repository completely would break things as we've all pulled / updated this
repo since the commit, so just removing it this time, and replacing with a text
file linking to datasheets.

git-svn-id: svn+ssh://forge.ecs.soton.ac.uk/projects/uavcamera@26 ea005762-10ae-0410-ac34-b80da5e0b649